Fix SD detect pin usage
SD detect pin was hardcoded in sd_conf.h while API provide a way to specify it.
This requires to provide the pin number instead of the GPIO_PIN_x pin and GPIOX port.
That's why sd_conf.h has been removed from the core.
Moreove several definition for the BSP are common or only dependent to SD
instance.
SD instance could be SDMMC1 or SDIO depending of the STM32 series.
Only the SD detect definition should be redefined thanks the variant.h
if the board has one.

+++ src/SD.cpp
@@ -81,7 +81,7 @@ uint8_t SDClass::begin()
   * @param  None
   * @retval TRUE or FALSE
   */
-uint8_t SDClass::begin(uint8_t cspin)
+uint8_t SDClass::begin(uint32_t cspin)
 {
 	/*##-1- Initializes SD IOs #############################################*/
 	if (_card.init(cspin)) {
