+++ src/Sd2Card.h
@@ -57,7 +57,7 @@ class Sd2Card {
  public:
 
   uint8_t init(void);
-  uint8_t init(uint8_t cspin);
+  uint8_t init(uint32_t cspin);
 
   /** Return the card type: SD V1, SD V2 or SDHC */
   uint8_t type(void) const;
